pyproject.toml
[project]
name = "llm-building-block"
version = "0.1.0"
description = ""
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
    "asyncio>=3.4.3",
    "beautifulsoup4>=4.13.4",
    "datasets>=3.6.0",
    "diskcache>=5.6.3",
    "llm>=0.26",
    "llm-anthropic>=0.17",
    "marimo>=0.14.10,<0.15.0",
    "matplotlib>=3.10.3",
    "migrate-to-uv (>=0.7.3,<0.8.0)",
    "mirascope[openai]>=1.25.4",
    "mohtml>=0.1.11",
    "mosync>=0.1.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "ruff>=0.12.2",
    "strip-tags>=0.6",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5,<7",
    "polars>=1.31.0,<2",
    "scikit-learn>=1.7.0,<2",
]

[tool.uv]
package = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
